+++ src/features/auth/redux/selectors.ts
@@ -2,3 +2,4 @@ import { CoreReduxState } from '@app/redux/types';
 
 export const selectAccessToken = ({ auth }: CoreReduxState) => auth.accessToken;
 export const selectIsLoadingAuth = ({ auth }: CoreReduxState) => auth.isLoading;
+export const selectUserInfo = ({ auth }: CoreReduxState) => auth.userInfo;