+++ src/features/wallet/components/HistoryTabContent/components/TransactionDetailsHeader/TransactionDetailsHeader.tsx
@@ -1,13 +1,19 @@
 import { AppText, AppView } from '@app/components';
 import { formatNumber } from '@app/utils/number';
 import { FC, useMemo } from 'react';
-import { UnionTransaction } from '@app/features/wallet/redux/types';
+import {
+  TransferTransaction,
+  UnionTransaction,
+} from '@app/features/wallet/redux/types';
 import { useAppTheme } from '@app/theme';
 import { TransactionType } from '@app/features/wallet/screens/Wallet/constants';
+import { useSelector } from 'react-redux';
+import { selectUserInfo } from '@app/features/auth/redux/selectors';
 
 export const TransactionDetailsHeader: FC<{ item: UnionTransaction }> = ({
   item,
 }) => {
+  const userInfo = useSelector(selectUserInfo);
   const { colors } = useAppTheme();
   const amount = useMemo(() => {
     switch (item.transactionType) {
@@ -16,14 +22,16 @@ export const TransactionDetailsHeader: FC<{ item: UnionTransaction }> = ({
       case TransactionType.Withdrawal:
         return -item.amount;
       case TransactionType.Transfer:
-        return -item.amount;
+        return userInfo?.id === (item as TransferTransaction).receiverUser.id
+          ? item.amount
+          : -item.amount;
       case TransactionType.Base:
         return item.amount;
       default:
         const _: never = item.transactionType;
         return _;
     }
-  }, [item.amount, item.transactionType]);
+  }, [item, userInfo?.id]);
 
   return (
     <AppView marginVertical={20} gap={10}>
@@ -31,7 +39,8 @@ export const TransactionDetailsHeader: FC<{ item: UnionTransaction }> = ({
         textAlign="center"
         textStyle="medium_26_32"
         color={amount > 0 ? colors.positiveStatus : colors.negativeStatus}>
-        {amount} {item.cryptoAsset.symbol}
+        {formatNumber(amount, undefined, 0, item.cryptoAsset.decimals ?? 0)}{' '}
+        {item.cryptoAsset.symbol}
       </AppText>
       <AppText
         textAlign="center"